(feat): Added projectKey to create_issue and list_issues. major refactor
